requirements.txt
numpy==1.17.0
matplotlib==3.1.2
opencv_python==4.1.2.30
torch==1.5.0
torchvision==0.6.0
tqdm==4.60.0
pillow==8.2.0